+++ NEWS.md
@@ -8,6 +8,8 @@ Added support for [Quarto](https://quarto.org), including:
 
 * Updated `stage_release()` to run `build_quarto_site()` when using Quarto
 
+* Updated `use_analysis_package()` to use Quarto for publishing <!--by default -->
+
 # rdev 1.7.2
 
 * Updated `.lintr` for `lintr` 3.1.1 (update `use_lintr()`)